// src/store/reducers/user/user.types.ts
export enum USERTYPES {
    USER_SIGNUP_START = 'user/USER_SIGNUP_START',
    USER_SIGNUP_SUCCESS = 'user/USER_SIGNUP_SUCCESS',
    USER_SIGNUP_FAILED = 'user/USER_SIGNUP_FAILED',
    USER_SIGNIN_START = 'user/USER_SIGNIN_START',
    USER_SIGNIN_SUCCESS = 'user/USER_SIGIN_SUCCESS',
    USER_SIGNIN_FAILED = 'user/USER_SIGNIN_FAILED',
    SET_CURRENT_USER = 'user/SET_CURRENT_USER',
    USER_SIGNOUT_START = 'user/USER_SIGNOUT_START',
    USER_SIGNOUT_SUCCESS = 'user/USER_SIGNOUT_SUCCESS',
    USER_SIGNOUT_FAILED = 'user/USER_SIGNOUT_FAILED',
}